Time: 3 ms (62.90%), Space: 6.7 MB (65.95%) - LeetHub

+++ 377-combination-sum-iv/377-combination-sum-iv.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int combinationSum4(vector<int>& a, int t) {
         int i,j,n=a.size(),m=1e9+7;
-        vector<unsigned long long> dp(t+1,0);
+        vector<unsigned int> dp(t+1,0);
         dp[0]=1;
         
         for(i=1;i<=t;i++)
